+++ sdk/cosmos/azure-cosmos/azure/cosmos/_cosmos_client_connection.py
@@ -170,6 +170,15 @@ class CosmosClientConnection:  # pylint: disable=too-many-public-methods,too-man
         self.availability_strategy: Union[CrossRegionHedgingStrategy, None] =\
             validate_client_hedging_strategy(availability_strategy)
         self.availability_strategy_executor: Optional[ThreadPoolExecutor] = availability_strategy_executor
+
+        # Mirror serving configuration (for per-request routing)
+        self._mirror_config = kwargs.pop('mirror_config', None)
+        if self._mirror_config is not None:
+            self._mirror_config = dict(self._mirror_config)
+        self._mirror_driver_client = None
+        self._mirror_driver_lock = threading.Lock()
+        self._mirror_driver_closed = False
+
         self.master_key: Optional[str] = None
 
         self.resource_tokens: Optional[Mapping[str, Any]] = None
@@ -296,6 +305,14 @@ class CosmosClientConnection:  # pylint: disable=too-many-public-methods,too-man
         # Use database_account if no consistency passed in to verify consistency level to be used
         self.session: Optional[_session.Session] = None
         self._set_client_consistency_level(database_account, consistency_level)
+
+    @property
+    def mirror_config(self):
+        """Fabric mirror configuration for per-request query routing.
+
+        :rtype: Optional[dict[str, Any]]
+        """
+        return self._mirror_config
 
     @property
     def _container_properties_cache(self) -> dict[str, dict[str, Any]]: